style: 💄 move `context/` outside
✅ Closes: #208
