+++ src/components/trade_page_body/trade_page_body.tsx
@@ -2,9 +2,9 @@ import {useContext} from 'react';
 import OrderSection from '../order_section/order_section';
 import Footer from '../footer/footer';
 import MarketSection from '../market_section/market_section';
-import {MarketProvider} from '../../lib/contexts/market_context';
-import {UserProvider} from '../../lib/contexts/user_context';
-import {GlobalContext, useGlobal} from '../../lib/contexts/global_context';
+import {MarketProvider} from '../../contexts/market_context';
+import {UserProvider} from '../../contexts/user_context';
+import {GlobalContext, useGlobal} from '../../contexts/global_context';
 import OrderSectionMobile from '../order_section_mobile/order_section_mobile';
 
 const TradePageBody = () => {
