+++ src/components/trading_chart_switch/trading_chart_switch.tsx
@@ -4,9 +4,9 @@ import {
   TRADING_CHART_SWITCH_BUTTON_SIZE,
 } from '../../constants/display';
 import Toggle from '../toggle/toggle';
-import {MarketContext} from '../../lib/contexts/market_context';
+import {MarketContext} from '../../contexts/market_context';
 import useWindowSize from '../../lib/hooks/use_window_size';
-import {useGlobal} from '../../lib/contexts/global_context';
+import {useGlobal} from '../../contexts/global_context';
 
 interface ITradingChartSwitchProps {
   getTradingViewType: (tradingViewState: string) => void;
