+++ src/contexts/global_context.tsx
@@ -1,11 +1,11 @@
 import {createContext, useState, useEffect, useContext, Dispatch, SetStateAction} from 'react';
-import useWindowSize from '../hooks/use_window_size';
-import {LAYOUT_BREAKPOINT} from '../../constants/display';
+import useWindowSize from '../lib/hooks/use_window_size';
+import {LAYOUT_BREAKPOINT} from '../constants/display';
 import {ToastContainer, toast as toastify} from 'react-toastify';
-import PositionDetailsModal from '../../components/position_details_modal/position_details_modal';
-import TransferProcessModal from '../../components/transfer_process_modal/transfer_process_modal';
+import PositionDetailsModal from '../components/position_details_modal/position_details_modal';
+import TransferProcessModal from '../components/transfer_process_modal/transfer_process_modal';
 import {MarketContext} from './market_context';
-import Toast from '../../components/toast/toast';
+import Toast from '../components/toast/toast';
 
 export interface IToastify {
   type: 'error' | 'warning' | 'info' | 'success';
